Safari-611.1.10/LayoutTests/crypto/workers/subtle/hmac-postMessage-worker-expected.txt

Test sending hmac crypto keys via postMessage to a worker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS All checks passed in worker
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'HMAC'
PASS key.algorithm.length is 128
PASS key.algorithm.hash.name is 'SHA-1'
PASS key.usages is ["sign", "verify"]
PASS bytesToASCIIString(exportedKey) is rawKeyAscii
PASS successfullyParsed is true

TEST COMPLETE
